requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
langchain
langchain-classic
langchain-openai
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib